refactor: better structure, types and exports
